+++ packages/store-api/src/platforms/vtex/clients/search/index.ts
