+++ packages/store-api/src/platforms/vtex/clients/search/types/ProductSearchResult.ts
